# Make close of ConstraintHandler 50x faster