While making #859 I discovered that `close!(ch)` was quite slow. Did a profile and turns out most time was spent on `copy!(ch.free_dofs, setdiff(1:ndofs(ch.dh), ch.prescribed_dofs))`. For the benchmarked test case this PR reduces the construction of a ConstraintHandler from about 2 s to 0.04 s. 
(Take exact speedup with grain of salt since done on laptop using battery power)

```julia
using Ferrite
grid = generate_grid(Quadrilateral, (1000, 1000))

dh = DofHandler(grid)
ip = Lagrange{RefQuadrilateral,2}()
add!(dh, :u, ip^2)
add!(dh, :v, ip)
close!(dh)

function make_ch(dh, grid)
    ch = ConstraintHandler(dh)
    add!(ch, Dirichlet(:u, getfaceset(grid, "right"), Returns(zero(Vec{2}))))
    add!(ch, Dirichlet(:v, getfaceset(grid, "left"), Returns(0.0)))
    close!(ch)
    return ch
end

ch = make_ch(dh, grid)
@time make_ch(dh, grid)

# PR
# run 1: 0.037628 seconds (4.26 k allocations: 93.643 MiB, 27.37% gc time)
# run 2: 0.027672 seconds (4.26 k allocations: 93.643 MiB)
  
# master 
# run 1: 1.989513 seconds (4.28 k allocations: 528.475 MiB, 2.04% gc time)
# run 2: 1.974789 seconds (4.28 k allocations: 528.475 MiB)
  
```