@fredrikekre 2023-12-11
Nice, perhaps extract to a `sorted_setdiff!` helper?