@KristofferC 2023-12-12
I wonder if there should exist a `free_dofs` variable computed at all? It seems wasteful to have a long array of numbers that is almost the dense set of 1:ndofs.