@KnutAM 2023-12-12
If storing it is a problem, perhaps a freedof iterator can be created using similar logic to this PR? Potentially storing a `Vector{UnitRange{Int}}` for efficiency?
It is quite nice to have an easy way of doing e.g. `r[fdofs]` (even if not needed as it is used in the examples)