@KristofferC 2023-12-12
> It is quite nice to have an easy way of doing e.g. r[fdofs] (even if not needed as it is used in the examples)

I guess I imagined something like https://github.com/JuliaData/InvertedIndices.jl where you would do `r[Not(cdofs)]`.