@KnutAM 2023-12-12
Looks very neat - should for sure give the required functionality!
Nice to add to an example and docs if someone feels like removing `ch.free_dofs`!